pdf-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
PyMuPDF==1.25.3
pymupdf4llm==0.1.8
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
